Disable DEFAULT_DISPLAY debug hook
Always target DEFAULT_EXTERNAL_DISPLAY for Maru Desktop.

This hook was previously used to show some immediate feedback that the
desktop system was running even on a faulty HDMI link. There have been
changes to the desktop settings that should make this obvious now so
this isn't needed.

Additionally, users can now start the desktop in the background, so
starting the desktop without a HDMI display connected is now a normal
use case.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>