mclient: fix crash when screen size changes
In the case where there is no valid display info returned from mflinger,
we can't ensure that the screen size is sync'd with the physical display
so we accept the change. Previously, this would cause mclient to crash
from a BadMatch error when calling XShmGetImage, but it was ok since the
systemd unit is set to reload and it would start up and re-initialize
itself fine off of the new screen size.

The problem was that we would neglect to free up the shmem region when
we crashed, potentially causing memory leaks in the long run. It's also
weird to just have mclient crash and restart.

We now have a custom error hander and some logic to clean-up and
reinitialize the shmem regions with X when the screen size changes.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>