mclient: improve screen resizing
If we can get the size of MDisplay (typically an HDMI display) from
mflinger, then we will aways sync the X server's size to that as usual.

But if we can't get the size of MDisplay (typically a virtual display on
the Android side), then accept any screen changes for X.

This lets the user resize the screen as needed when they are connected
to a virtual display with unknown size.

This also includes some error handling improvements and a CRITICAL log
level for conditions where the program must terminate or risk unknown
behavior.

Signed-off-by: Preetam D'Souza <preetamjdsouza@gmail.com>